Fix warnings that got introduced by changes from the debian people.

+++ exif.c
@@ -477,7 +477,7 @@ static void ProcessExifDir(unsigned char * DirStart, unsigned char * OffsetBase,
             }
         }
         if (DumpExifMap){
-            printf("Map: %05ld-%05ld: Directory\n",DirStart-OffsetBase, DirEnd+4-OffsetBase);
+            printf("Map: %05ld-%05ld: Directory\n",(int)(DirStart-OffsetBase), (int)(DirEnd+4-OffsetBase));
         }
